Unit tests functional
* Let's see how much is broken for integration tests
