+++ dockstore-webservice/src/main/java/io/dockstore/webservice/resources/proposedGA4GH/ToolsApiExtendedServiceImpl.java
@@ -46,7 +46,6 @@ import io.dockstore.webservice.core.WorkflowVersion;
 import io.dockstore.webservice.helpers.PublicStateManager;
 import io.dockstore.webservice.jdbi.ToolDAO;
 import io.dockstore.webservice.jdbi.WorkflowDAO;
-import io.swagger.api.impl.ToolsApiServiceImpl;
 import io.swagger.api.impl.ToolsImplCommon;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpHost;
@@ -106,8 +105,8 @@ public class ToolsApiExtendedServiceImpl extends ToolsExtendedApiService {
      * @param organization
      * @return
      */
-    private List<Entry> getPublishedByOrganization(String organization) {
-        final List<Entry> published = new ArrayList<>();
+    private List<Entry<?, ?>> getPublishedByOrganization(String organization) {
+        final List<Entry<?, ?>> published = new ArrayList<>();
         published.addAll(workflowDAO.findPublishedByOrganization(organization));
         published.addAll(toolDAO.findPublishedByNamespace(organization));
         published.sort(Comparator.comparing(Entry::getGitUrl));
@@ -119,12 +118,12 @@ public class ToolsApiExtendedServiceImpl extends ToolsExtendedApiService {
         return Response.ok().entity(getPublishedByOrganization(organization)).build();
     }
 
-    private List<io.swagger.model.Tool> workflowOrgGetList(String organization) {
+    private List<io.openapi.model.Tool> workflowOrgGetList(String organization) {
         List<Workflow> published = workflowDAO.findPublishedByOrganization(organization);
         return published.stream().map(c -> ToolsImplCommon.convertEntryToTool(c, config)).collect(Collectors.toList());
     }
 
-    private List<io.swagger.model.Tool> entriesOrgGetList(String organization) {
+    private List<io.openapi.model.Tool> entriesOrgGetList(String organization) {
         List<Tool> published = toolDAO.findPublishedByNamespace(organization);
         return published.stream().map(c -> ToolsImplCommon.convertEntryToTool(c, config)).collect(Collectors.toList());
     }
@@ -224,10 +223,10 @@ public class ToolsApiExtendedServiceImpl extends ToolsExtendedApiService {
     public Response setSourceFileMetadata(String type, String id, String versionId, String platform, String platformVersion, String relativePath, Boolean verified,
         String metadata) {
 
-        ToolsApiServiceImpl impl = new ToolsApiServiceImpl();
-        ToolsApiServiceImpl.ParsedRegistryID parsedID = new ToolsApiServiceImpl.ParsedRegistryID(id);
-        Entry entry = impl.getEntry(parsedID, Optional.empty());
-        Optional<? extends Version> versionOptional;
+        io.openapi.api.impl.ToolsApiServiceImpl impl = new io.openapi.api.impl.ToolsApiServiceImpl();
+        io.openapi.api.impl.ToolsApiServiceImpl.ParsedRegistryID parsedID = new io.openapi.api.impl.ToolsApiServiceImpl.ParsedRegistryID(id);
+        Entry<?, ?> entry = impl.getEntry(parsedID, Optional.empty());
+        Optional<? extends Version<?>> versionOptional;
 
         if (entry instanceof Workflow) {
             Workflow workflow = (Workflow)entry;
@@ -241,7 +240,7 @@ public class ToolsApiExtendedServiceImpl extends ToolsExtendedApiService {
             return Response.status(Response.Status.NOT_FOUND).build();
         }
         if (versionOptional.isPresent()) {
-            Version version = versionOptional.get();
+            Version<?> version = versionOptional.get();
             // so in this stream we need to standardize relative to the main descriptor
             Optional<SourceFile> correctSourceFile = impl
                 .lookForFilePath(version.getSourceFiles(), relativePath, version.getWorkingDirectory());
