+++ dockstore-webservice/src/main/java/io/swagger/api/impl/ApiV2BetaVersionConverter.java
@@ -1,0 +1,219 @@
+/*
+ *    Copyright 2017 OICR
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package io.swagger.api.impl;
+
+import java.lang.reflect.InvocationTargetException;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import javax.ws.rs.core.Response;
+
+import com.google.common.collect.Lists;
+import io.dockstore.common.DescriptorLanguage;
+import io.openapi.model.Checksum;
+import io.openapi.model.ImageData;
+import io.swagger.model.DescriptorType;
+import io.swagger.model.ExtendedFileWrapper;
+import io.swagger.model.FileWrapper;
+import io.swagger.model.Metadata;
+import io.swagger.model.MetadataV1;
+import io.swagger.model.Tool;
+import io.swagger.model.ToolVersion;
+import org.apache.commons.beanutils.BeanUtils;
+import org.apache.commons.beanutils.ConvertUtils;
+import org.apache.commons.beanutils.Converter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * Converts between the V2-beta version of the GA4GH TRS to V1
+ *
+ * @author gluu, dyuen
+ * @since 21/12/17
+ */
+public final class ApiV2BetaVersionConverter {
+    private static final Logger LOG = LoggerFactory.getLogger(ApiV2BetaVersionConverter.class);
+
+    static {
+        ConvertUtils.register(new ToolClassConverter(), io.swagger.model.ToolClass.class);
+        ConvertUtils.register(new DescriptorTypeConverter(), io.swagger.model.DescriptorType.class);
+    }
+
+    private ApiV2BetaVersionConverter() {
+        // utility class
+    }
+
+    public static Response convertToVersion(Response response) {
+        Object object = response.getEntity();
+        if (object instanceof List) {
+            List<Object> arrayList = (List<Object>)object;
+            List<Object> newArrayList = new ArrayList<>();
+            for (Object innerObject : arrayList) {
+                if (innerObject instanceof io.openapi.model.Tool) {
+                    io.openapi.model.Tool tool = (io.openapi.model.Tool)innerObject;
+                    newArrayList.add(getTool(tool));
+                } else if (innerObject instanceof io.openapi.model.ToolVersion) {
+                    io.openapi.model.ToolVersion toolVersion = (io.openapi.model.ToolVersion)innerObject;
+                    newArrayList.add(getToolVersion(toolVersion));
+                } else if (innerObject instanceof ExtendedFileWrapper) {
+                    // TODO: this probably is not the case in v2 beta, check it
+                    // v1 annoying expects a 1 Dockerfile list to be returned unwrapped
+                    Object extendedWrapperConverted = getWrapper((ExtendedFileWrapper)innerObject);
+                    if (arrayList.size() == 1
+                        && ((ExtendedFileWrapper)innerObject).getOriginalFile().getType() == DescriptorLanguage.FileType.DOCKERFILE) {
+                        return getResponse(extendedWrapperConverted);
+                    }
+                    newArrayList.add(extendedWrapperConverted);
+                } else {
+                    newArrayList.add(innerObject);
+                }
+            }
+            return getResponse(newArrayList);
+        } else if (object instanceof io.openapi.model.ToolVersion) {
+            io.openapi.model.ToolVersion toolVersion = (io.openapi.model.ToolVersion)object;
+            ToolVersion betaToolVersion = getToolVersion(toolVersion);
+            return getResponse(betaToolVersion);
+        } else if (object instanceof io.openapi.model.Tool) {
+            io.openapi.model.Tool tool = (io.openapi.model.Tool)object;
+            Tool betaTool = getTool(tool);
+            return getResponse(betaTool);
+        } else if (object instanceof Metadata) {
+            Metadata metadata = (Metadata)object;
+            MetadataV1 metadataV1 = new MetadataV1(metadata);
+            return getResponse(metadataV1);
+        } else if (object instanceof io.openapi.model.FileWrapper) {
+            if (object instanceof ExtendedFileWrapper) {
+                return getResponse(getWrapper((ExtendedFileWrapper)object));
+            }
+            return getResponse(object);
+        }
+        return response;
+    }
+
+    public static Tool getTool(io.openapi.model.Tool tool) {
+        Tool betaTool = new Tool();
+        try {
+            BeanUtils.copyProperties(betaTool, tool);
+            betaTool.setToolname(tool.getName());
+            betaTool.setHasChecker(tool.isHasChecker());
+            Set<String> authors = new HashSet<>();
+            tool.getVersions().stream().map(io.openapi.model.ToolVersion::getAuthor).filter(Objects::nonNull).forEach(authors::addAll);
+            if (authors.isEmpty()) {
+                betaTool.setAuthor("Unknown author");
+            } else {
+                betaTool.setAuthor(String.join("", authors));
+            }
+
+            betaTool.setSigned(false);
+            betaTool.setContains(Lists.newArrayList());
+            // convert versions now
+            betaTool.setVersions(new ArrayList<>());
+            for (io.openapi.model.ToolVersion version : tool.getVersions()) {
+                ToolVersion oldVersion = getToolVersion(version);
+                betaTool.getVersions().add(oldVersion);
+            }
+            betaTool.setVerified(tool.getVersions().stream().anyMatch(io.openapi.model.ToolVersion::isVerified));
+            Set<String> sources = new HashSet<>();
+            tool.getVersions().stream().map(io.openapi.model.ToolVersion::getVerifiedSource).forEach(sources::addAll);
+            betaTool.setVerifiedSource(sources.isEmpty() ? "[]" : sources.stream().collect(Collectors.joining("\",\"", "[\"", "\"]")));
+        } catch (IllegalAccessException | InvocationTargetException e) {
+            LOG.error("unable to backwards convert tool");
+            throw new RuntimeException(e);
+        }
+        return betaTool;
+    }
+
+    public static ToolVersion getToolVersion(io.openapi.model.ToolVersion toolVersion) {
+        ToolVersion betaToolVersion = new ToolVersion();
+        try {
+            BeanUtils.copyProperties(betaToolVersion, toolVersion);
+            // look like it has issues converting enums
+            betaToolVersion.setDescriptorType(Lists.newArrayList());
+            toolVersion.getDescriptorType().forEach(type -> {
+                betaToolVersion.getDescriptorType().add(DescriptorType.fromValue(type.name()));
+            });
+
+            // looks like BeanUtils has issues due to https://issues.apache.org/jira/browse/BEANUTILS-321 and https://github.com/swagger-api/swagger-codegen/issues/7764
+            betaToolVersion.setVerified(toolVersion.isVerified());
+            betaToolVersion.setVerifiedSource(toolVersion.getVerifiedSource().isEmpty() ? "[]"
+                : toolVersion.getVerifiedSource().stream().collect(Collectors.joining("\",\"", "[\"", "\"]")));
+            betaToolVersion.setContainerfile(toolVersion.isContainerfile());
+            betaToolVersion.setImageName(
+                toolVersion.getImages().stream().filter(Objects::nonNull).map(ImageData::getImageName).collect(Collectors.joining()));
+            // this is a bit weird, but seems to be current behaviour, also need to get rid of the double lambda
+            final Optional<Optional<String>> first = toolVersion.getImages().stream().filter(Objects::nonNull).map(
+                item -> item.getChecksum().stream().filter(check -> check.getType().equals(ToolsImplCommon.DOCKSTORE_IMAGEID))
+                    .map(Checksum::getChecksum).findFirst()).findFirst();
+            if (first.isPresent() && first.get().isPresent()) {
+                betaToolVersion.setImage(first.get().get());
+            } else {
+                betaToolVersion.setImage("");
+            }
+            betaToolVersion.setRegistryUrl(
+                toolVersion.getImages().stream().filter(Objects::nonNull).map(ImageData::getRegistryHost).collect(Collectors.joining()));
+        } catch (IllegalAccessException | InvocationTargetException e) {
+            LOG.error("unable to backwards convert toolVersion");
+            throw new RuntimeException(e);
+        }
+        return betaToolVersion;
+    }
+
+    public static FileWrapper getWrapper(io.openapi.model.FileWrapper wrapper) {
+        FileWrapper oldWrapper = new FileWrapper();
+        oldWrapper.setContent(wrapper.getContent());
+        oldWrapper.setUrl(wrapper.getUrl());
+        return oldWrapper;
+    }
+
+    public static FileWrapper getWrapper(ExtendedFileWrapper wrapper) {
+        FileWrapper oldWrapper = new FileWrapper();
+        oldWrapper.setContent(wrapper.getContent());
+        oldWrapper.setUrl(wrapper.getUrl());
+        return oldWrapper;
+    }
+
+    private static Response getResponse(Object object) {
+        Response.ResponseBuilder responseBuilder = Response.ok(object);
+        return responseBuilder.build();
+    }
+
+    public static class DescriptorTypeConverter implements Converter {
+        @Override
+        public <T> T convert(Class<T> type, Object value) {
+            io.swagger.model.DescriptorType betaToolClass = io.swagger.model.DescriptorType.fromValue(value.toString());
+            return (T)betaToolClass;
+        }
+    }
+
+    public static class ToolClassConverter implements Converter {
+        @Override
+        public <T> T convert(Class<T> type, Object value) {
+            io.swagger.model.ToolClass betaToolClass = new io.swagger.model.ToolClass();
+            try {
+                BeanUtils.copyProperties(betaToolClass, value);
+            } catch (IllegalAccessException | InvocationTargetException e) {
+                LOG.error("unable to backwards convert ToolClass");
+                throw new RuntimeException(e);
+            }
+            return (T)betaToolClass;
+        }
+    }
+}
