+++ dockstore-webservice/src/main/java/io/swagger/model/ExtendedFileWrapper.java
@@ -19,6 +19,7 @@ import java.util.Objects;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import io.dockstore.webservice.core.SourceFile;
+import io.openapi.model.FileWrapper;
 
 /**
  * Used to store additional transient information about files to be returned from the GA4GH endpoints
